Refactor create validations to OO
